App Improvements

+++ app/src/main/java/com/shegs/artreasurehunt/ui/arena/CreateArenaDialog.kt
@@ -10,7 +10,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
@@ -28,7 +27,6 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.input.ImeAction
@@ -191,7 +189,7 @@ fun CreateArenaDialog(
                                     id = UUID.randomUUID().toString(),
                                     arenaName = arenaName,
                                     arenaDesc = arenaDescription,
-                                    arenaLocation = arenaLocation,
+                                    arenaLocation = arenaLocation.toDouble(),
                                     imageResId = randomImageResId // Provide a default image
                                 )
                                 viewModel.createArena(newArena)
